+++ pkg/android/find-android-manifest.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"path/filepath"
 
+	"github.com/bugsnag/bugsnag-cli/pkg/log"
 	"github.com/bugsnag/bugsnag-cli/pkg/utils"
 )
 
@@ -18,11 +19,11 @@ import (
 //   - appBuildPath: The root build directory where Android build outputs are located.
 //   - variant: The build variant (e.g., "debug", "release") whose manifest paths
 //     should be searched.
+//   - logger: logger used to emit debug output when manifest is not found.
 //
 // Returns:
 //   - string: The resolved manifest path if found, otherwise an empty string.
-//   - error: Non-nil if the manifest cannot be found.
-func FindAndroidManifest(appBuildPath string, variant string) (string, error) {
+func FindAndroidManifest(appBuildPath string, variant string, logger log.Logger) string {
 	var err error
 
 	mergedManifestPath := filepath.Join(appBuildPath, "intermediates", "merged_manifests")
@@ -32,7 +33,8 @@ func FindAndroidManifest(appBuildPath string, variant string) (string, error) {
 	}
 
 	if err != nil {
-		return "", err
+		logger.Info("No AndroidManifest.xml located: a single variant directory couldn't be found")
+		return ""
 	}
 
 	primaryAppManifestPath := filepath.Join(mergedManifestPath, variant, "AndroidManifest.xml")
@@ -43,9 +45,13 @@ func FindAndroidManifest(appBuildPath string, variant string) (string, error) {
 
 	for _, path := range paths {
 		if utils.FileExists(path) {
-			return path, nil
+			logger.Debug(fmt.Sprintf("AndroidManifest.xml located at: %s", path))
+			return path
+		} else {
+			logger.Debug(fmt.Sprintf("AndroidManifest.xml not found at: %s", path))
 		}
 	}
 
-	return "", fmt.Errorf("unable to locate AndroidManifest.xml for variant %s", variant)
+	logger.Info(fmt.Sprintf("No AndroidManifest.xml located for variant %s", variant))
+	return ""
 }